remvoe isProjectKit. What is this?
HTML exports have a CORS issue that happens when you try to access the window.parent element
This is because the export is on a different domain to the export screen where the preview is shown
Why do we check for a projectKit. What is a project kit? it's not a preview state I know of in the platform and I see no reason why this would be a reason to not run the validation
I also removed the preventExportOverflow flag. If you don't want the limiters to run on export then only call the functions if the state is not equal to "export"