added custom bleed and trim support
trimMarks and placeholders were not able to be updated without reloading the page until now
Plus you can set custom amounts of bleed since it's now css driven. We no longer need JS to have crop marks. The boilerplate still takes care of it for backwards compatibly but all you need for bleed is the CSS file, the crop-marks html and to set the CSS variables of trim and bleed